# Make map() and broadcast() faster for Union{T, Missing/Nothing} element types