@nanosoldier 2018-02-03
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/25828#issuecomment-362767021):
```
ProcessExitedException()
```
cc @ararslan