@ararslan 2018-02-04
That was after upgrading to the more recent HTTP version, which was supposed to help the intermittent ignoring, but instead it segfaults... So I've pinned at an older HTTP version for now.

@nanosoldier `runbenchmarks(ALL, vs=":master")`