@nanosoldier 2018-02-04
[Your benchmark job](https://github.com/JuliaLang/julia/pull/25828#issuecomment-362870623) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/50f9b9c1e725906299f165156f3b480e8b8f0088/fe0e478_vs_97a1856/report.md). cc @ararslan